Create README.md
# ProspectivityMapping_TL
Code Source for the work - Quantifying Uncertainty and Improving Prospectivity Mapping in Mineral Belts Using Transfer Learning and Random Forest: A Case Study of Copper Mineralization in the Superior Craton Province, Quebec, Canada.
Link to paper: xxx

It is possible due to a missing seed value that the result be slightly different from the one present in the paper. However, the conclusion is still the same.